+++ packages/api/endpoints/schemas.js
@@ -11,9 +11,7 @@ function get(event, cb) {
 }
 
 function handler(event, context) {
-  return handle(event, context, true, (cb) => {
-    return get(event, cb);
-  });
+  return handle(event, context, true, (cb) => get(event, cb));
 }
 
 module.exports = handler;