+++ packages/integration-tests/index.js
@@ -226,18 +226,24 @@ async function addCollections(stackName, bucketName, dataDirectory) {
 }
 
 /**
- * add providers to database
+ * add providers to database.
  *
  * @param {string} stackName - Cloud formation stack name
  * @param {string} bucketName - S3 internal bucket name
  * @param {string} dataDirectory - the directory of provider json files
+ * @param {string} s3Host - bucket name to be used as the provider host for 
+ * S3 providers. This will override the host from the seed data. Defaults to null, 
+ * meaning no override.
  * @returns {Promise.<integer>} number of providers added
  */
-async function addProviders(stackName, bucketName, dataDirectory) {
+async function addProviders(stackName, bucketName, dataDirectory, s3Host = null) {
   const providers = await setupSeedData(stackName, bucketName, dataDirectory);
 
   const promises = providers.map((provider) => limit(() => {
     const p = new Provider();
+    if(s3Host && provider.protocol === 's3') {
+      provider.host = s3Host;
+    }
     console.log(`adding provider ${provider.id}`);
     return p.delete({ id: provider.id }).then(() => p.create(provider));
   }));